+++ src/spanner/healthcheck.ts
@@ -1,38 +1,30 @@
+import { BaseHealthIndicatorService } from '@causa/runtime/nestjs';
 import { Database } from '@google-cloud/spanner';
 import { Injectable } from '@nestjs/common';
-import {
-  HealthCheckError,
-  HealthIndicator,
-  HealthIndicatorResult,
-} from '@nestjs/terminus';
+import { HealthCheckError, HealthIndicatorResult } from '@nestjs/terminus';
 
 /**
  * The key used to identify the Spanner health indicator.
  */
-const SPANNER_HEALTH_KEY = 'spanner';
+const SPANNER_HEALTH_KEY = 'google.spanner';
 
 /**
  * A service testing the availability of the Spanner service.
  */
 @Injectable()
-export class SpannerHealthIndicator extends HealthIndicator {
+export class SpannerHealthIndicator extends BaseHealthIndicatorService {
   constructor(private readonly database: Database) {
     super();
   }
 
-  /**
-   * Checks the health of the Spanner service by running a dummy query (`SELECT 1`).
-   *
-   * @returns The health of the Spanner service.
-   */
-  async isHealthy(): Promise<HealthIndicatorResult> {
+  async check(): Promise<HealthIndicatorResult> {
     try {
       await this.database.run('SELECT 1');
 
       return this.getStatus(SPANNER_HEALTH_KEY, true);
     } catch (error: any) {
       throw new HealthCheckError(
-        'Failed to run healthcheck Spanner query.',
+        'Failed to check health by running Spanner query.',
         this.getStatus(SPANNER_HEALTH_KEY, false, { error: error.message }),
       );
     }